fix `malloc: double free for ptr` crashes
