+++ Sources/MarkdownView/Renderer/Renderer+CodeBlock.swift
@@ -1,8 +1,5 @@
 import Markdown
 import SwiftUI
-#if os(iOS) || os(macOS)
-import Highlightr
-#endif
 
 // MARK: - Inline Code Block
 extension Renderer {
@@ -42,135 +39,3 @@ extension Renderer {
         }
     }
 }
-
-#if os(iOS) || os(macOS)
-struct HighlightedCodeBlock: View {
-    var language: String?
-    var code: String
-    var theme: CodeHighlighterTheme
-    
-    @Environment(\.markdownFont) private var font
-    @Environment(\.colorScheme) private var colorScheme
-    @State private var attributedCode: AttributedString?
-    @State private var showCopyButton = false
-    
-    private var id: String {
-        "\(colorScheme) mode: " + (language ?? "No Language Name") + code
-    }
-
-    var body: some View {
-        Group {
-            if let attributedCode {
-                SwiftUI.Text(attributedCode)
-            } else {
-                SwiftUI.Text(code)
-            }
-        }
-        .task(id: id, highlight)
-        .lineSpacing(5)
-        .font(font.codeBlock)
-        .padding()
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .background(.quaternary.opacity(0.5), in: RoundedRectangle(cornerRadius: 8))
-        .gesture(
-            TapGesture()
-                .onEnded { _ in showCopyButton.toggle() }
-        )
-        .overlay(alignment: .topTrailing) {
-            if showCopyButton {
-                CopyButton(content: code)
-                    .padding(8)
-                    .transition(.opacity.animation(.easeInOut))
-            }
-        }
-        .overlay(alignment: .bottomTrailing) {
-            codeLanguage
-        }
-        .onHover { showCopyButton = $0 }
-    }
-    
-    @ViewBuilder
-    private var codeLanguage: some View {
-        if let language {
-            SwiftUI.Text(language.uppercased())
-                .font(.callout)
-                .padding(8)
-                .foregroundStyle(.secondary)
-        }
-    }
-    
-    @Sendable private func highlight() {
-        guard let highlighter = Highlightr.shared else { return }
-        highlighter.setTheme(to: colorScheme == .dark ? theme.darkModeThemeName : theme.lightModeThemeName)
-        let language = highlighter.supportedLanguages().first(where: { $0.lowercased() == self.language?.lowercased() })
-        if let highlightedCode = highlighter.highlight(code, as: language) {
-            withAnimation {
-                attributedCode = AttributedString(highlightedCode)
-            }
-        }
-    }
-}
-
-extension Highlightr {
-    static var shared: Highlightr? = Highlightr()
-}
-
-struct CopyButton: View {
-    var content: String
-    @State private var copied = false
-    #if os(macOS)
-    @ScaledMetric private var size = 12
-    #else
-    @ScaledMetric private var size = 18
-    #endif
-    @State private var isHovering = false
-    
-    var body: some View {
-        Button(action: copy) {
-            Group {
-                if copied {
-                    Image(systemName: "checkmark")
-                        .transition(.opacity.combined(with: .scale))
-                } else {
-                    Image(systemName: "doc.on.clipboard")
-                        .transition(.opacity.combined(with: .scale))
-                }
-            }
-            .font(.system(size: size))
-            .frame(width: size, height: size)
-            .padding(8)
-            .contentShape(Rectangle())
-        }
-        .foregroundStyle(.primary)
-        .background(
-            .quaternary.opacity(0.2),
-            in: RoundedRectangle(cornerRadius: 5, style: .continuous)
-        )
-        .overlay {
-            RoundedRectangle(cornerRadius: 5, style: .continuous)
-                .stroke(.quaternary, lineWidth: 1)
-        }
-        .brightness(isHovering ? 0.3 : 0)
-        .buttonStyle(.borderless) // Only use `.borderless` can behave correctly when text selection is enabled.
-        .onHover { isHovering = $0 }
-    }
-    
-    private func copy() {
-        #if os(macOS)
-        NSPasteboard.general.clearContents()
-        NSPasteboard.general.setString(content, forType: .string)
-        #else
-        UIPasteboard.general.string = content
-        #endif
-        Task {
-            withAnimation(.spring()) {
-                copied = true
-            }
-            try await Task.sleep(nanoseconds: 2_000_000_000)
-            withAnimation(.spring()) {
-                copied = false
-            }
-        }
-    }
-}
-#endif
